control: separate files, remove TODO
